Reorganize files, remove deprecated/unused code
