+++ .gitignore
@@ -135,9 +135,6 @@ dmypy.json
 config.json
 config/*
 !config/.gitkeep
-user/*
-!user/.gitkeep
-modules/anime/config/alID.json
 assets/img_gen/*.jpg
 !skel/.env
 main.log*
