+++ modules/core/events.py
@@ -2,8 +2,6 @@ import discord, json, os, string, re
 from os import path
 
 from modules.core.client import Client
-from modules.config.config import Config
-from modules.config.user import User
 from modules.core.resources import Resources
 
 bot = Client.bot
@@ -28,36 +26,12 @@ class Events:
 			serverID = str(guild.id)
 			serverName = str(guild.name)
 			if path.exists(os.getcwd()+"/config/"+serverID+".json"):
-				Config.cfgUpdate(serverID, "Name", serverName)
+				Resources.config.cfgUpdate(serverID, "Name", serverName)
 			else:
 				with open(os.getcwd()+"/config/"+serverID+".json", "x") as server_json:
 					json_data = {"Name": serverName}
 					json.dump(json_data, server_json)
 
-		if not path.exists(os.getcwd()+"/modules/anime/config/alID.json"):
-			with open(os.getcwd()+"/modules/anime/config/alID.json", "x") as al_json:
-				json_data = {0: 0}
-				json.dump(json_data, al_json)
-
-
-		for user in bot.users:
-			userID = str(user.id)
-			userName = str(user.name)
-			if path.exists(os.getcwd()+"/user/"+userID+".json"):
-				User.userUpdate(userID, "Name", userName)
-			else:
-				with open(os.getcwd()+"/user/"+userID+".json", "x") as user_json:
-					json_data = {"Name": userName}
-					json.dump(json_data, user_json)
-
-			# for AL Auto-updates
-			alID = User.userRead(str(user.id), "alID")
-			if alID!=0 or None:
-				with open(os.getcwd()+"/modules/anime/config/alID.json", 'r') as al_json:
-					json_data = json.load(al_json)
-					json_data[str(user.id)] = alID
-				with open(os.getcwd()+"/modules/anime/config/alID.json", 'w') as al_json:
-					al_json.write(json.dumps(json_data))
 
 	@bot.event
 	async def on_member_join(member):
@@ -67,10 +41,10 @@ class Events:
 					role = discord.utils.get(guild.roles, name='RecWell')
 					await discord.Member.add_roles(member, role)
 					break
-		if Config.cfgRead(str(member.guild.id), "welcomeOn")==True:
-			welcomeMsg = Config.cfgRead(str(member.guild.id), "welcomeMsg")
+		if Resources.config.cfgRead(str(member.guild.id), "welcomeOn")==True:
+			welcomeMsg = Resources.config.cfgRead(str(member.guild.id), "welcomeMsg")
 			welcomeMsgFormatted = welcomeMsg.format(member=member.mention)
-			await bot.get_channel(Config.cfgRead(str(member.guild.id), "welcomeChannel")).send(welcomeMsgFormatted)
+			await bot.get_channel(Resources.config.cfgRead(str(member.guild.id), "welcomeChannel")).send(welcomeMsgFormatted)
 
 	def determine_reaction(msg, reactions):
 		for reaction in reactions:
