+++ modules/core/resources/config.py
