+++ modules/queries/anime/anilist.py
