+++ modules/queries/anime/anilist2.py
