+++ modules/queries/anime/mal.py
