+++ modules/queries/anime/safebooru.py
