+++ modules/queries/music/search.py
